src/encoding/image_encoder.py
"""
Image Encoder — BioMedCLIP.

Encodes medical images (chest X-rays) into the same representation
space as clinical text, enabling cross-modal retrieval and
multimodal grounding.
"""

import logging

import numpy as np
from PIL import Image

logger = logging.getLogger(__name__)


class ImageEncoder:
    """BioMedCLIP-based medical image encoder.

    Encodes chest X-ray images into embeddings that can be aligned
    with text embeddings for cross-modal retrieval and grounding.

    Parameters
    ----------
    model_name : str
        HuggingFace model ID for BioMedCLIP.
    device : str
        Device to run inference on ('cpu' or 'cuda').
    """

    def __init__(
        self,
        model_name: str = "microsoft/BiomedCLIP-PubMedBERT_256-vit_base_patch16_224",
        device: str = "cpu",
    ):
        self.model_name = model_name
        self.device = device
        self.model = None
        self.preprocessor = None

    def load_model(self) -> None:
        """Load the BioMedCLIP model and preprocessor."""
        # TODO: Implement model loading
        raise NotImplementedError

    def encode(
        self, images: list[Image.Image], batch_size: int = 16
    ) -> np.ndarray:
        """Encode a batch of medical images into embeddings.

        Parameters
        ----------
        images : list[Image.Image]
            Input medical images.
        batch_size : int
            Batch size for encoding.

        Returns
        -------
        np.ndarray
            Embedding matrix of shape (len(images), embedding_dim).
        """
        # TODO: Implement image encoding
        raise NotImplementedError
