requirements.txt
# ============================================================
# Requirements — Hinglish Grounded Medical RAG
# ============================================================

# ---------- Core ML / Deep Learning ----------
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.30.0
datasets>=2.19.0

# ---------- Text Encoding (LaBSE) ----------
sentence-transformers>=3.0.0

# ---------- Image Encoding (BioMedCLIP) ----------
open_clip_torch>=2.24.0

# ---------- Vector Database (FAISS) ----------
faiss-cpu>=1.8.0

# ---------- Fine-Tuning (QLoRA / DPO) ----------
peft>=0.11.0
trl>=0.9.0
bitsandbytes>=0.43.0

# ---------- Data Processing ----------
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
Pillow>=10.3.0

# ---------- Medical Data ----------
pydicom>=2.4.0

# ---------- NLP & Evaluation ----------
nltk>=3.8.0
rouge-score>=0.1.2
evaluate>=0.4.0

# ---------- Visualization ----------
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.22.0

# ---------- Experiment Tracking ----------
wandb>=0.17.0

# ---------- Application Interface ----------
streamlit>=1.35.0

# ---------- Configuration ----------
PyYAML>=6.0.0
python-dotenv>=1.0.0

# ---------- Utilities ----------
tqdm>=4.66.0
requests>=2.32.0
huggingface_hub>=0.23.0
